requirements.txt
azure-functions>=1.18.0,<2.0.0
requests>=2.31.0,<3.0.0
certifi>=2024.0.0
